TASK-2: Add Abstract Entity
