+++ src/main/java/ru/example/projectmanagement/entities/AbstractEntity.java
@@ -1,0 +1,19 @@
+package ru.example.projectmanagement.entities;
+
+import javax.persistence.*;
+
+@MappedSuperclass
+public abstract class AbstractEntity {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    private Long id;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+}
