+++ src/main/java/ru/example/projectmanagement/entities/Project.java
@@ -8,12 +8,7 @@ import java.util.List;
 
 @Entity
 @Table(name = "project")
-public class Project {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id")
-    private Long id;
+public class Project extends AbstractEntity {
 
     @Column(name = "project_name")
     private String name;
@@ -42,14 +37,6 @@ public class Project {
     private List<Task> tasks;
 
     public Project() {
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
     }
 
     public String getName() {
